DB 통하여 애플리캐이션 기능 구현

+++ src/main/kotlin/com/group/libraryapp/dto/book/response/BookStatResponse.kt
@@ -4,5 +4,5 @@ import com.group.libraryapp.domain.book.BookType
 
 data class BookStatResponse (
     val type: BookType,
-    var count: Int,
+    var count: Long,
 )
