+++ src/test/kotlin/com/group/libraryapp/service/book/BookServiceTest.kt
@@ -135,17 +135,17 @@ class BookServiceTest @Autowired constructor(
 
         // then
         assertThat(results).hasSize(2)
-        assertCount(results,BookType.COMPUTER, 2)
+        assertCount(results,BookType.COMPUTER, 2L)
         // val computerDto = results.first { results -> results.type == BookType.COMPUTER }
         // assertThat(computerDto.count).isEqualTo(2)
 
-        assertCount(results,BookType.SCIENCE, 1)
+        assertCount(results,BookType.SCIENCE, 1L)
         // val scienceDto = results.first { results -> results.type == BookType.SCIENCE }
         // assertThat(scienceDto.count).isEqualTo(1)
     }
 
     // 조회 캐이스도 private 로 요약할 수 있음
-    private fun assertCount(results: List<BookStatResponse>, type: BookType, count: Int){
+    private fun assertCount(results: List<BookStatResponse>, type: BookType, count: Long){
         assertThat(results.first{ result -> result.type == type }.count).isEqualTo(count)
     }
 }
